Create header library for goldfish includes
In order to be able to use the goldfish includes such as the qemu pipe
includes in a project that uses soong the includes must be in a header
library. Soong only allows include paths that are contained in
subdirectories of the project directory. Since these headers are outside
of all other projects it needs to be inclued as a dependency instead.

BUG: 114288761
Test: Compile test project with Android.bp file
Change-Id: I3ccb2befc986b7bf1f2f6dff3a5812c861618a77