Make DHCP server assign interface-based addresses
Currently the DHCP server used in the router namespace assigns addresses
as configured on the command line. This is redundant information as all
the necessary information can be deduced from the current interface
address anyway.

Remove most of the command line options in favor of assigning addresses
from the same subnet that the interface is in and a gateway that is the
same address as the interface.

BUG: 111997081
Test: Run emulator and verify that radio and wifi networking works
Change-Id: I8b6adb76f7445b7650da8f2a59b020b4fef991ee