Create a network manager for the emulator
Create a network manager for the emulator that will contain
functionality related to network setup and management. Initial
functionality includes address assignments to wifi interfaces as they
come and go, and receiving and acting on wifi related commands over the
qemu pipe. Currently the network manager only supports creation of
access points and blocking/unblocking network traffic for an access
point.

In order to have a hostapd configuration file that can change the setup
scripts will now copy the template configuration file to the vendor data
partition which is writable and then modify it.

BUG: 111997081
Test: Run CTS Verifier test Multinetwork connectivity
Change-Id: I754e0c74cbaea9daa130915bed47ef3b2fba164e