Make netmgr support WiFi direct by forwarding
In order to support WiFi Direct we need two devices to be able to send
raw WiFi frames between each other. This is to emulate the situation
where the WiFi radios in the devices are able to talk to each other
directly without any access point or established networks.

The way this is done is by reading WiFi frames from the monitor
interface provided by the virtual WiFi driver and then transporting them
to another emulator. On the other emulator the frames are then injected
into the virtual WiFi driver's monitor interface.

The frames are read from the monitor interface and then transported out
of Android using the QEMU pipe. Similarly data from the other emulator
is read from the QEMU pipe and then injected on the monitor interface.

BUG: 111996306
Test: Run CTS Verifier WiFi Direct tests
Change-Id: I4f8abf483776ae97a1b6184d012b3a23e55e87a2