Copy verified_boot.xml unconditionally.
The BOARD_AVB_ENABLE setting is not available at this stage of the build
unless a vendor file sets it.  As
make/target/board/BoardConfigGsiCommon.mk is setting the flag it could
be argued that setting it in vendor files as well could be redundant or contradictory.

For the emulator, copying the permission seems to have no ill effect,
even if we boot without verified boot enabled.

Bug: b/118443541
Test:
  - Build image and asserted that xml file is  present.
  - Started emulator both with and without verified boot enabled
  - Ran CtsSecurityTestCases
android.security.cts.VerifiedBootTest#testVerifiedBootSupport and
asserted that it is now passing.

Change-Id: I66b749a6b4712ae5fedd290dfc6d8dc00e5c6c1b