Set ro.control_privapp_permissions=enforce.
This is primarily needed for CtsPermission2TestCases
android.permission2.cts.PrivappPermissionsTest#testPrivappPermissionsEnforcement
to pass.

Also, we need to change the permission package in Q from
com.android.sdksetup to com.google.android.sdksetup

I decided to do this change in master first (not AOSP) to limit the
scope of unanticipated issues with the permission change.

Bug: b/118443972
Test: Asserted that the CTS test passes.  Did a full CTS run and
asserted that the basic pass numbers are in the correct ballpark.

Change-Id: I3cb028e7008bb224698bac0a20d83d623e36195a