Fix kernel BPF network statistics
In the past Android has maintained its own kernel code for network
statistics per process, called xt_qtaguid. This is now being phased out
in favor of BPF, a flexible system to attach custom code to certain
points in the kernel at runtime. This is used to track network
statistics per process instead to avoid having to maintain non-upstream
kernel code.

With kernel 4.14 the emulator needs to switch to using BPF. The kernel
had the right configuration but the system image was having SELinux
denials that prevented the setup and the memlock memory limit was not
sufficient to fit the network statistics.

Fix the emulators local netd.te SELinux policy to allow netd to create,
read and write the BPF maps used for statistics. Set the memlock memory
limit to the minimum 8 MB as required by
https://source.android.com/devices/tech/datausage/ebpf-traffic-monitor

Test: atest CtsNetTestCases
BUG: 123351370
Change-Id: Ibd56bec09906ba167079774adcfe99b4510e6f23