Partial revert of "Fix kernel BPF network statistics"
This reverts sepolicy changes of commit dab4e8bc8bff124a2cbc345c1254b73669885988.

These were bogus, netd selinux policy is not device specific
and this is now getting in the way of adding neverallow rules.

No other device has such rules, so goldfish does not need them either.
And even if it does, this is *NOT* the right place for them.

Test: treehugger will
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I52d7b012bf0515e667d372ef8a31da721b18a913