Pretend to support packet offloading in WiFi HAL
Currently Android uses packet offloading to support TCP keepalive
packets while allowing the CPU to sleep. The keepalive packet is handed
off to the WiFi hardware with an instruction to retransmit at a certain
frequency. The emulator does not support this which caused a new CTS
test to fail. Before this change the emulator WiFi HAL would just drop
the packet anyway saying the operation is not supported.

This change pretends to support packet offloading in the WiFi HAL which
allows the CTS tests to pass. The keepalive packet is still dropped but
the HAL indicates that it was successful.

BUG: 134773926, 134774684, 130897952
Test: CtsNetTestCases (testSocketKeepaliveUnprivileged, testSocketKeepaliveLimit, testCreateTcpKeepalive)
Change-Id: I67ddad49f279374e237ba59871bf7c5fb16a3c18