[hw-codec] Move media profile configs into data partition.
This will enable us to dynamically change the media_codecs_google_video.xml being
used, so we can run tests on different profiles.

If the emulator does not support this configurable
media_codecs_google_video.xml
mode, it will simply fall back to the
media_codecs_google_video_default.xml, which is the one we currently are
using at the moment.

Test: Manual.
1) Boot emulator without DynamicMediaProfile feature support. There
should be a symbolic link from
/data/vendor/etc/media_codecs_google_video_default.xml ->
/data/vendor/etc/media_codecs_google_video.xml

2) Boot emulator with DynamicMediaProfile feature support, but with
display resolution < 1080p (1920x1080). The result should be the same as 1).

3) Boot emulator as in 2), but with display resolution > 1080p.
The symbolic link should now be
/data/vendor/etc/media_codecs_google_video_v2.xml ->
/data/vendor/etc/media_codecs_google_video.xml
Change-Id: I17befc1949bc509cf4f45c7ebd29696f37d4f48b