New implementation for WiFi Direct forwarding
With the recent change to use bridging for emulator networking and WiFi
support the previos WiFi Direct implementation stopped working. This
happened because WiFi frames are now too large to inject on the monitor
interface. The MTU of the AP wireless interface can no longer be lowered
to create smaller frames that allow injection headers. The bridge
requires all interfaces to have the same MTU and the emulator requires
an MTU of 1500.

To work around this this change uses a feature of the mac80211_hwsim
driver that allows userspace full control over the flow of WiFi frames
through the driver. This way a userspace forwarder reads all frames
entering the kernel driver and then ensures that they are routed to
their intended destination. This also includes remote destinations such
as another emulator.

Additionally this allows acknowledgement of frames from remote sources
which increases reliability and the forwarder can therefore also
implement retransmissions. This fixes flakiness issues with the previous
solution of WiFi Direct that appeared when running the CTS Verifier.

BUG: 120612311
Test: Manually run CTS Verifier WiFi Direct tests
Change-Id: I3670434eaa28b7e139746172b65833cd404ab0da
Merged-In: I3670434eaa28b7e139746172b65833cd404ab0da