goldfish: Replace manifest-arm.xml with a more flexible way of excluding specific HALs.
The manifest-arm.xml file was added as a way of excluding the gnss HAL from
the ARM build, but over time it diverged from manifest.xml and in fact can
no longer be used to boot. (It also ended up picking up the gnss HAL, even
though the whole point of adding the file was to not include it.)

Instead of maintaining multiple xml files here with mostly the same contents,
just maintain a single manifest.xml together with manifest.$HAL.xml files with
manifests for specific HALs, and use assemble_vintf to stitch them together.

Bug: 142352330
Change-Id: If286228b54ca8dd359e8ce4cc869ee7c8f29d4b4