goldfish: Add lunch targets for the ARM Fixed Virtual Platform.
The Armv8-A Base Platform FVP is a simulator for ARM systems. The FVP provides
access to the newest ARM architectural features before they are available
in silicon, so it may be used as a platform for early development of uses
of new features in the Android platform.

The fvp-eng and fvp-userdebug lunch targets may be used to produce
a system image that is suitable to be used with the FVP.

Bug: 142352330
Change-Id: I5dafa3bbb8c20d4d41cc6aa97f6797e0c1376df1