fvp: Switch to the new "automatic" GKI kernel module loader.
In the GKI world, the way that kernel modules are supposed to be loaded
is by including them in the initramfs and letting first-stage init load
them. Switch to that for FVP, which allows us to remove a bunch of custom
logic for loading kernel modules.

Normally, the bootloader is supposed to concatenate the Android initramfs
and the initramfs for the kernel modules and let the kernel combine
them. However, the bootloader that we're using with FVP (U-Boot) doesn't
support concatenation, so we implement it in the build system.

Also, update the kernel build instructions for MTE, which now requires
a newer binutils.

Bug: 142352330
Change-Id: I1cc8d141f8524bfc01a7d4ca972c50454d5d70a3