fvp: Set PRODUCT_SHIPPING_API_LEVEL to 29.
r.android.com/1207843 caused the configstore service to no longer be installed
by default if PRODUCT_SHIPPING_API_LEVEL was unset, which led to a failure
to boot. This highlighted the fact that we should probably be setting it. The
variable should be set to as high a value as possible, in order to be as close
as possible to the upcoming hardware that the fvp target is meant to prototype.

Unfortunately setting it to 30 (and removing configstore from manifest.xml)
does not work yet because of remaining hard dependencies on configstore
(see b/124531214#comment43), so set it to 29 instead.

Bug: 142352330
Bug: 124531214
Change-Id: I3cb7559fd43e36ad3db9fa7925523b9a5f2b0b6b