radio: uses modem simulator from cuttlefish
This adds ModemSimulator=on feature flag to indicate
that modem simulator (from cuttlefish) is used, and
host side should use the corresponding backend.

The guest rild uses virtual port to communicate
with the host backend. Corresponding selinux rules
are also setup to allow that access.

Compared to the previous implementation, the current
one allows more customization, since operator, iccard
content are stored in the iccprofile_for_sim0.xml
and it is distributed together with system images
into the data folder.

Note: this depends on the emulator host support
go/oag/1314949.

BUG: 155797805

Change-Id: I80cbc71cdb7d1fcd541e8be4ad875726ddd3e97e