codecs: update codec xml file
1. remove obsolete decoder OMX.google.goldfish.*
as they render inside guest and are no longer necessary.
2. update max concurrent instances for OMX.android.goldfish.*
to 4, as that is a realistic limit on host.
3. update performance data, the CTS DecoderPerformanceTest
are picky on the accuracy of those numbers, and it appears
that the tolerance is about 10%. the data is from my local
xeon(Intel(R) Xeon(R) CPU E5-2690 v4 @ 2.60GHz, nvidia M2000),
other machine could vary a bit.

BUG: 160616123

Test: atest android.media.cts.VideoDecoderPerfTest#testAvc*
Test: atest android.media.cts.VideoDecoderPerfTest#testVp*
Change-Id: Iee1ab4700a2b4ea1e89ae29a57140f394fc5f3d6
Merged-In: Iee1ab4700a2b4ea1e89ae29a57140f394fc5f3d6