Address jitter in tinyalsa pcm_read
AudioRecordTest#testTimestamp expects jitter
to be below 6ms while pcm_read happens to be
around 20ms. This change moves pcm_read
into a separate thread and adds a ring buffer
to read audio data. When the audio data is not
available it is returned as zeroes.

Test: atest android.media.cts.AudioRecordTest
Test: x86_64 CtsMediaTestCases: Passed: 29, Failed: 0
Bug: 161581762
Change-Id: Ic86788beafba11d6f25054a261d324bf2558161a
Merged-In: Ic86788beafba11d6f25054a261d324bf2558161a
Signed-off-by: Roman Kiryanov <rkir@google.com>