Add libdrm.so to same_process_hal_file context list.
A recent change to SwiftShader [1] caused it to start depending on
libdrm. Therefore we need to give libdrm.so the same_process_hal_file
context so that it can be opened by processes that need to load
the SwiftShader shared library which is an SP-HAL.

While here remove references to non-existent files (these files live
in the egl directory which already receives the same_process_hal_file
context via system/sepolicy/private/file_contexts).

[1] https://swiftshader-review.googlesource.com/c/SwiftShader/+/48508

Bug: 142352330
Change-Id: I6602d1b5ed57683fde02aba14c4476f4764ee938