Do not mount /data before late-fs.
Currently the emulator attempts to mount /data during the "fs" trigger.
Since "fs" comes before "post-fs", we haven't yet run restorecon on
/metadata, or created /metadata/vold. As a result, vold fails to land
bits necessary for metadata-encryption.

Because vold performs a recursive mkdir, it incidentally tries to create
/metadata/vold when populating the key directory. Since the emulator
hasn't run restorecon yet, /metadata has the wrong label (cache_file
instead of metadata_file), and a misleading audit message appears. The
emulator works around this by including a device-specific rule for vold.

If we attempt to restorecon /metadata earlier in bootup (as is needed
for b/173476209), the problem becomes apparent: vold is denied access to
the _correct_ label.

The fix is to make sure that only early partitions are mounted in "fs",
and /data is marked as "latemount" and only mounted in "late-fs". This
matches other devices which use metadata encryption.

Bug: 173476209
Test: emulator boots with early restorecon
Change-Id: I4b0ca48bd5bf6a218a445ac81a0bcc899418d96e