fvp: qemu: Move setprops to early-init.
It turns out that boot may be too late for the
android.hardware.graphics.allocator@2.0-service process to see the
correct value of the ro.hardware.gralloc system property, since this
process is launched at around the same time. As a result we may end
up loading the default gralloc HAL into this process, resulting in
HAL mismatches with other processes. Move the setprops to early-init
to avoid this problem.

Bug: 178231152
Change-Id: Ib0b5fc2fc64dac46fd2357a28b4beea89a6baa42