radio: update iccprofile to cuttlefish
This is to bring in the SimPhonebook fixes
from cuttlefish.

now the profile is almost the same as cf,
except the phone service provider
where goldfish has been using tmoble.

TODO: it would be nicer to split
the common parts into one file and
the different ones to the minimum.

Bug: 190537169
Change-Id: I980929ae21e189ce7ab326abb20860b8572f506b