goldfish: add shortened 64bit boards
There is this problem of "build fingerprint too long", and
this is to mitigate a bit.

This will replace the verbose version of the following boards

emulator64_x86_64 -> emu64x
emulator64_x86_64_arm64 -> emu64xa
emulator64_arm64 -> emu64a

Right now, the content are exactly the same. next step
is to replace verbose named board with shortened one.

Change-Id: Ie736b01f7e0fb566e3f31a6004d0f232dc5c41b4