fvp: Switch to specific overlay
https://r.android.com/1861493 broke the network on fvp, because fvp
only has one functional network adapter, eth0. Since there is no way
to change this just for fvp, and to avoid further breakage, let's
create a cut-down overlay specific to fvp.

Most peripherals are absent from fvp, so there is no need to
replicate the vast majority of the configuration files from
goldfish. We do keep a few customisations in SettingsProvider to
make sure the virtual screen is always on.

Test: boot fvp-eng, check that the network is functional
Change-Id: Ifb24b4060e467d89fbdc3d5ba30621817e6579fd