fvp: Remove most hardware permissions
Since https://r.android.com/1818094, fvp does not provide Bluetooth
support. This is mostly fine, except that that
handheld_core_hardware.xml still claims that the
android.hardware.bluetooth permissions is provided. This causes
com.android.bluetooth to crash with the following asertion:

  bluetooth: assertion 'bt_hci_ != nullptr' failed

It is always restarted, causing the crash to happen continuously.

Since fvp does not actually provide most of the features listed in
handheld_core_hardware.xml, it feels that the best thing to do is to
stop using it. It turns out we just need one permission from that
file to boot without issue, that is android.software.app_widgets,
and luckily there is already a standard file that we can use to
provide it.

Without some of the permissions in handheld_core_hardware.xml, it is
likely that many CTS tests (and potentially other system tests) will
fail, but it is anyway unlikely they would pass on the FVP platform.

Change-Id: Ied0184b06497d3a90b799f3f8d4cab349e756633