Report wlan0 and wlan1 to the wifi hal depending on presence
The wifi HAL needs to know about all available wifi interfaces, and can
get confused when one of the relevant interfaces is not initially
reported but used later.

This is relevant on cuttlefish because it defines both wlan0 and wlan1
interfaces. I believe one is intended to be 2.4 ghz and the other is 5
ghz though I am not completely sure.

The Info::Info() wifi code now checks `/sys/class/net` for wlan0 and
wlan1 to decide which names to report.

Test: atest command from b/263758606
Bug: b/263758606
Change-Id: I84364321da1ac64a1b57279c1d32f8538f4d37e3