add pixel_fold overlay
This is for newer implementation of foldable
that uses two built-in displays. one for
unfolded and one for folded.

This should not impact phone emulator, as
the changes are all inside the pixel_fold overlay
which is only enabled for pixel_fold device.

Bug: 289373564
Test: presubmit
Change-Id: I3262389cef82f0492d3f027194a6870e02ff6ce5
Merged-In: I3262389cef82f0492d3f027194a6870e02ff6ce5