Mount system_dlkm in ext4 format
To test the generic kernel, the system_dlkm must be replaced
with the image containing compatible kernel modules. The
system_dlkm image is available in only ext4 format in kernel
branches. This commit adds an entry to fstab.ranchu.arm. If
the device cannot mount system_dlkm as erofs, it falls back
to ext4.

Bug: 307432651
Test: export ANDROID_PRODUCT_OUT=~/acloud_gf_4/image/arm64 \
      ANDROID_TMP=~/acloud_gf_4/instance \
      ANDROID_BUILD_TOP=~/acloud_gf_4/instance ; \
      acloud_gf_4/emulator/emulator/emulator -verbose \
      -show-kernel -read-only -ports 5560,5561 -no-window \
      -kernel acloud_gf_4/kernel \
      -ramdisk acloud_gf_4/mixed_ramdisk \
      -qemu -append androidboot.verifiedbootstate=orange
Change-Id: I5f8e2fbd321ff9ef7f36438e2f1e0829f748dd53