radio: relax RadioNetwork::validateNetworkScanRequest
was too restrictive.

The NetworkScanApiTest#testRequestNetworkScan* failed
because the RadioNetwork::startNetworkScan method
returned RadioError::INVALID_ARGUMENTS for empty
RadioAccessSpecifier::bands in one of
NetworkScanRequest::specifiers.

I don't know if having no bands in a specifier
is a correct thing, but this is what the tests
send and rejecting it makes the tests unhappy.

Bug: 450621746
Test: atest CtsCarrierApiTestCases
Change-Id: I16dddc38f93e91aaa310021477e4d8900025a36d
Signed-off-by: Roman Kiryanov <rkir@google.com>