Add emu_camera_main json file
We are working on refactoring the camera2 emulator HAL to have configurable number of cameras. In that process, we have added this emu_camera_main file which is included as part of apex. Cuttlefish is using apex file and hence the issue was not reproducible on cuttlefish. Since goldfish is not using apex,we are seeing the issue in goldfish

Bug: 454444185
Test: test goldfish
Change-Id: Id4261f7cf41244cd5d027cf5e486ce2e7a5da63d